you can't leave a guild you own

+++ socket/event/leaveguild.go
@@ -27,6 +27,17 @@ func OnLeaveGuild(ws *globals.Client, rawMap map[string]interface{}, limiter *ra
 	if globals.Guilds[data.Guild] == nil || globals.Guilds[data.Guild].Clients[userid] == nil {
 		return
 	}
+	var ownerID  string
+	err = harmonydb.DBInst.QueryRow("SELECT owner FROM guilds WHERE guildid=$1", data.Guild).Scan(&ownerID)
+	if err != nil {
+		golog.Warnf("Error leaving guild : %v", err)
+		sendErr(ws, "We were unable to get you to leave the guild for some reason")
+		return
+	}
+	if ownerID == userid {
+		sendErr(ws, "You cannot leave a guild you own")
+		return
+	}
 	_, err = harmonydb.DBInst.Exec("DELETE FROM guildmembers WHERE userid=$1 AND guildid=$2", userid, data.Guild)
 	// GUILD STUCK! GUILD STUCK! PLEASE! I BEG YOU!
 	if err != nil {
